fix for ConnectionTest, GP-6600
